- Adding logic to make use of PDAnalyst's simulation data

+++ Apromore-Custom-Plugins/Process-Simulation-Info-Logic/src/main/java/org/apromore/processsimulation/model/Resource.java
@@ -43,5 +43,5 @@ public class Resource {
     @XmlAttribute
     private String timetableId;
     @XmlAttribute
-    private int totalAmount;
+    private long totalAmount;
 }
